+++ src/main/java/com/revolvingmadness/sculk/language/builtins/classes/instances/FunctionInstance.java
@@ -2,10 +2,12 @@ package com.revolvingmadness.sculk.language.builtins.classes.instances;
 
 import com.revolvingmadness.sculk.Sculk;
 import com.revolvingmadness.sculk.gamerules.SculkGamerules;
+import com.revolvingmadness.sculk.language.Argument;
 import com.revolvingmadness.sculk.language.ErrorHolder;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinClass;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinFunction;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinType;
+import com.revolvingmadness.sculk.language.errors.TypeError;
 import com.revolvingmadness.sculk.language.interpreter.Interpreter;
 import com.revolvingmadness.sculk.language.interpreter.errors.MaxArgumentError;
 import com.revolvingmadness.sculk.language.interpreter.errors.Return;
@@ -13,15 +15,16 @@ import com.revolvingmadness.sculk.language.lexer.TokenType;
 import com.revolvingmadness.sculk.language.parser.nodes.statement_nodes.StatementNode;
 
 import java.util.List;
+import java.util.ListIterator;
 import java.util.Objects;
 
 public class FunctionInstance extends BuiltinFunction {
-    public final List<String> arguments;
+    public final List<Argument> arguments;
     public final List<StatementNode> body;
     public final String name;
     public final BuiltinType returnType;
 
-    public FunctionInstance(String name, List<String> arguments, BuiltinType returnType, List<StatementNode> body) {
+    public FunctionInstance(String name, List<Argument> arguments, BuiltinType returnType, List<StatementNode> body) {
         this.name = name;
         this.arguments = arguments;
         this.returnType = returnType;
@@ -42,12 +45,21 @@ public class FunctionInstance extends BuiltinFunction {
             throw ErrorHolder.invalidArgumentCount(this.name, this.arguments.size(), arguments.size());
         }
 
-        int argumentNumber = 0;
+        ListIterator<Argument> argumentIterator = this.arguments.listIterator();
 
-        for (String argumentName : this.arguments) {
-            BuiltinClass argumentValue = arguments.get(argumentNumber);
-            interpreter.variableTable.declare(List.of(TokenType.CONST), argumentName, argumentValue);
-            argumentNumber++;
+        while (argumentIterator.hasNext()) {
+            Argument argument = argumentIterator.next();
+            BuiltinClass typeClass = interpreter.variableTable.getOrThrow(argument.type).value;
+            BuiltinClass value = arguments.get(argumentIterator.previousIndex());
+
+            if (!(typeClass instanceof BuiltinType type)) {
+                throw new TypeError("The type of an argument cannot be an instance");
+            }
+
+            /* TODO do what you did with functions but with methods */
+            /* TODO */
+
+            interpreter.variableTable.declare(List.of(TokenType.CONST), type, argument.name, value);
         }
 
         try {