+++ src/main/java/com/revolvingmadness/sculk/language/builtins/functions/AbsFunction.java
@@ -7,6 +7,7 @@ import com.revolvingmadness.sculk.language.builtins.classes.instances.FloatInsta
 import com.revolvingmadness.sculk.language.builtins.classes.instances.IntegerInstance;
 import com.revolvingmadness.sculk.language.builtins.classes.types.FloatType;
 import com.revolvingmadness.sculk.language.builtins.classes.types.IntegerType;
+import com.revolvingmadness.sculk.language.builtins.classes.types.NumberType;
 import com.revolvingmadness.sculk.language.interpreter.Interpreter;
 
 import java.util.List;
@@ -14,20 +15,18 @@ import java.util.List;
 public class AbsFunction extends BuiltinFunction {
     @Override
     public BuiltinClass call(Interpreter interpreter, List<BuiltinClass> arguments) {
-        if (arguments.size() != 1) {
-            throw ErrorHolder.invalidArgumentCount("abs", 1, arguments.size());
+        this.validate("abs", arguments, List.of(NumberType.TYPE));
+
+        BuiltinClass number = arguments.get(0);
+
+        if (number.instanceOf(IntegerType.TYPE)) {
+            return new IntegerInstance(Math.abs(number.toInteger()));
         }
 
-        BuiltinClass object = arguments.get(0);
-
-        if (object.instanceOf(IntegerType.TYPE)) {
-            return new IntegerInstance(Math.abs(object.toInteger()));
+        if (number.instanceOf(FloatType.TYPE)) {
+            return new FloatInstance(Math.abs(number.toFloat()));
         }
 
-        if (object.instanceOf(FloatType.TYPE)) {
-            return new FloatInstance(Math.abs(object.toFloat()));
-        }
-
-        throw ErrorHolder.argumentRequiresType(1, "abs", IntegerType.TYPE, object.getType());
+        throw ErrorHolder.argumentRequiresType(1, "abs", NumberType.TYPE, number.getType());
     }
 }