+++ src/main/java/com/revolvingmadness/sculk/language/builtins/functions/FloorFunction.java
@@ -3,10 +3,10 @@ package com.revolvingmadness.sculk.language.builtins.functions;
 import com.revolvingmadness.sculk.language.ErrorHolder;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinClass;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinFunction;
-import com.revolvingmadness.sculk.language.builtins.classes.instances.IntegerInstance;
+import com.revolvingmadness.sculk.language.builtins.classes.instances.FloatInstance;
 import com.revolvingmadness.sculk.language.builtins.classes.types.FloatType;
 import com.revolvingmadness.sculk.language.builtins.classes.types.IntegerType;
-import com.revolvingmadness.sculk.language.errors.TypeError;
+import com.revolvingmadness.sculk.language.builtins.classes.types.NumberType;
 import com.revolvingmadness.sculk.language.interpreter.Interpreter;
 
 import java.util.List;
@@ -14,20 +14,18 @@ import java.util.List;
 public class FloorFunction extends BuiltinFunction {
     @Override
     public BuiltinClass call(Interpreter interpreter, List<BuiltinClass> arguments) {
-        if (arguments.size() != 1) {
-            throw ErrorHolder.invalidArgumentCount("floor", 1, arguments.size());
+        this.validate("floor", arguments, List.of(NumberType.TYPE));
+
+        BuiltinClass number = arguments.get(0);
+
+        if (number.instanceOf(IntegerType.TYPE)) {
+            return new FloatInstance(Math.floor(number.toInteger()));
         }
 
-        BuiltinClass object = arguments.get(0);
-
-        if (object.instanceOf(IntegerType.TYPE)) {
-            return new IntegerInstance((long) Math.floor(object.toInteger()));
+        if (number.instanceOf(FloatType.TYPE)) {
+            return new FloatInstance(Math.floor(number.toFloat()));
         }
 
-        if (object.instanceOf(FloatType.TYPE)) {
-            return new IntegerInstance((long) Math.floor(object.toFloat()));
-        }
-
-        throw new TypeError("Function 'floor' requires integer or float");
+        throw ErrorHolder.argumentRequiresType(1, "floor", NumberType.TYPE, number.getType());
     }
 }