+++ src/main/java/com/revolvingmadness/sculk/language/builtins/functions/TypeFunction.java
@@ -1,10 +1,10 @@
 package com.revolvingmadness.sculk.language.builtins.functions;
 
-import com.revolvingmadness.sculk.language.ErrorHolder;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinClass;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinFunction;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinType;
 import com.revolvingmadness.sculk.language.builtins.classes.instances.StringInstance;
+import com.revolvingmadness.sculk.language.builtins.classes.types.ObjectType;
 import com.revolvingmadness.sculk.language.interpreter.Interpreter;
 
 import java.util.List;
@@ -12,9 +12,7 @@ import java.util.List;
 public class TypeFunction extends BuiltinFunction {
     @Override
     public BuiltinClass call(Interpreter interpreter, List<BuiltinClass> arguments) {
-        if (arguments.size() != 1) {
-            throw ErrorHolder.invalidArgumentCount("type", 1, arguments.size());
-        }
+        this.validate("type", arguments, List.of(ObjectType.TYPE));
 
         BuiltinClass object = arguments.get(0);
         BuiltinType type = object.getType();